requirements.txt
fastapi
uvicorn
sqlmodel
supabase-py
pytest
